Fix CSS SSR
The server was generating different MUI classes to the client.

The fix is to explicitly generate the MUI classes on the server outside
of the markup the client knows about, and then remove the classes when
the client renders for the first time.

Also improves the API for glamor SSR.